cue: support optional field lookup in LookupPath
This is done by allowing Selectors to be converted
to an optional form (if they aren't already).

It also supports the use of the Selectors AnyIndex and
AnyString for getting `T` for  `[string]: T` and `[...T]`.

This also allows deprecating Elem and Template, which
will be done in a separate CL.

Change-Id: I6074382f12259c3dd87557471d80f82720a84779