cue/build: remove support for file lists
This is a breaking change! We do not have "go fix"
support because:
1) it is anticipated that these weren't really used by users
2) they are easy to fix by hand, yet somewhat tricky to
do with Go fix.

Users are expected to use the lists of type []*File.

Tool or test files can be found in IgnoredFiles, if the
Test or Tool option were not used.

Use Instance.RelPath to get the relative path of a file.

Change-Id: Ib430fce26562c584129ac9183ebf72ccd1466054