internal/core/adt: improve performance for repeated calls to Unify/Fill
This is done by skipping SpawnRef for computed vertices.

Repeatedly applying unification caused a new closedInfo to
be inserted at the node's position, resulting in a growing
list to process when calling Accept (called by Unify, which
is called by Fill). This caused processing time to become
quadratic for that node.

There are more gains to be made, but this should be a big
help already.

Issue #899

Change-Id: I96ec77bb4bd09b69b9da023e0fb2b8ed5c297080