cue: reimplement Decode to not use MarshalJSON
The roundtrip to JSON resulted in a few incompatibilities

Note that this copies the logic in encoding/json for
parsing json tags and disambiguating names.

Code that previously did a JSON marshal and relied
on that detecting errors before modifying Go values
must now explicitly check for concreteness first.

Fixes #922

Change-Id: I9a55151163183ab84d1269d51f180127869a9a15