ci: trigger unity and cuelang.org builds/equivalent on tip/new versions
A new tip merge of CUE should:

* trigger a unity build
* trigger a redeploy of tip.cuelang.org

A new release of CUE should:

* trigger a unity build against that version
* prompt us to release a new version of cuelang.org

We could choose to make the release of a new version of cuelang.org on a
new version of CUE automatic (or automate the creation of a PR that
releases a new version of cuelang.org), but for now we simply trigger a
build which will fail if the current version of CUE referenced by
cuelang.org is _not_ the latest. This in effect will be the prompt we
need to perform a manual release of cuelang.org, using the latest CUE.

This change reflects the desired end state described above. Automating
the release of cuelang.org/creation of a PR can follow later.

Change-Id: I6623fd50c6c15e4f423bc204d65e3dc98569da7a