cue: add InferBuiltins EncodeOption
This functionality is currently partially internal or deprecated.
This adds this functionality to the new API.

This also removes a cyclic dependency in package dependencies
that needs to be removed for the implementation of structure
sharing.

This moves resolveExpr to below BuildExpr, as their bodies are
almost the same (functionally the same at the moment).
This also removes dead code.

Expresses the old CompileExpr in terms of the new BuildExpr.
This is mostly done as a  verification and to allow automatic
rewrites in the future.

Change-Id: I7d63aa097dd8f652b9b0b0181a2b148c8bb1f2d4