cue: prepare API implementation for structure sharing
The CUE API relies on the Parent field to reflect the
path down which the user descended to obtain the
value. With structure sharing, this means this parent
field may vary by context.

This now adds another field to Value, which keeps
track of a differing parent Value chain. The field
only needs to be set in the case of actual structure
sharing.

Change-Id: Ia03de7d249ce72e72104740ce649827e7b07f2ad