encoding/openapi: finalise when constructing openapi AST
When building OpenAPI enums from a CUE instance, specifically one that
contains an enum imported from protobuf, we don't care about the
definition-encoded integer enum value (introduced as part of dcfff000).
Instead we only care about the final value of the enum.

Prior to dcfff000, an imported protobuf enum would look like this:

    #CaptureMode: "DEFAULT" | "IPTABLES" | "NONE"

Post dcfff000, it looks like:

    #CaptureMode: {
            "DEFAULT"
            #enumValue: 0
    } | {
            "IPTABLES"
            #enumValue: 1
    } | {
            "NONE"
            #enumValue: 2
    }

Therefore, in the process of building an enum we simply need to finalise
the CUE Value prior to deriving its syntax.

The test as part of this CL comes from a real Istio schema that
uncovered this bug in the first place.

Fixes #977

Change-Id: I5b4cbb2f35b0535488d8f77b1e2166f89ed2644d