ci: fix up new version trigger
Currently, a new version release results in a failure to trigger a build
against unity for that new version, due to a poorly formed JSON payload.
This is due to fragile shell script that is hand-written as part of the
CUE workflow specification. This hand-written shell script must be
tested manually (for now) and we insufficiently tested before.

Fixed and tested.

Change-Id: I7b165042ffda9bdceb84e3faab775f5d2711527e